+++ src/render/uniform_binding.test.ts
@@ -7,8 +7,11 @@ import {
     Uniform2f,
     Uniform3f,
     Uniform4f,
+    UniformFloatArray,
+    UniformColorArray,
     UniformMatrix4f
 } from './uniform_binding';
+import {Color} from '@maplibre/maplibre-gl-style-spec';
 
 describe('Uniform Binding', () => {
     test('Uniform1i', () => {
@@ -118,6 +121,42 @@ describe('Uniform Binding', () => {
         expect(u.current).toEqual(ident);
         u.set(ident);
         u.set([2, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1]);
+    });
+
+    test('UniformColorArray', () => {
+        expect.assertions(4);
+
+        const context = {
+            gl: {
+                uniform4fv: () => { expect(true).toBeTruthy(); }
+            }
+        } as any as Context;
+
+        const u = new UniformColorArray(context, 0);
+        expect(u.current).toEqual(new Array<Color>);
+        const v = [new Color(0.1, 0.2, 0.3), new Color(0.7, 0.8, 0.9)];
+        u.set(v);
+        expect(u.current).toEqual(v);
+        u.set(v);
+        u.set([new Color(0.3, 0.4, 0.5), new Color(0.4, 0.5, 0.6), new Color(0.5, 0.6, 0.7)]);
+    });
+
+    test('UniformFloatArray', () => {
+        expect.assertions(4);
+
+        const context = {
+            gl: {
+                uniform1fv: () => { expect(true).toBeTruthy(); }
+            }
+        } as any as Context;
+
+        const u = new UniformFloatArray(context, 0);
+        expect(u.current).toEqual(new Array<number>);
+        const v = [1.2, 3.4];
+        u.set(v);
+        expect(u.current).toEqual(v);
+        u.set(v);
+        u.set([5.6, 7.8, 9.1]);
     });
 
 });