+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 🐞 Bug fixes
 
 -   Fix `queryRenderedFeatures` on globe view when crossing international date line ([#5765](https://github.com/maplibre/maplibre-gl-js/pull/5765))
+-   Fix globe `unproject` to clamp points to horizon ([#5771](https://github.com/maplibre/maplibre-gl-js/pull/5771))
 - _...Add new stuff here..._
 
 ## 5.3.1